refactor(live): extract turnCycleGuard to its own file
Issue #15 names turn_cycle_guard.go as the implementation file for the
guard state machine. The PR inlined the type, methods, and
applyTurnCycleGuard in base_flow_live.go (already 975 lines spanning
sender/receiver/processMessage). The existing turn_cycle_guard_test.go
referenced symbols from a sibling file, which is awkward to navigate.

Pure move — no behavior change. The struct, its three methods
(onModelContent, onTurnComplete, reset), and applyTurnCycleGuard now
live in internal/llminternal/turn_cycle_guard.go alongside their tests.
Subsequent fixes (Interrupted-resets, function-response-resets) edit
this file directly.
